Make the memory reporting multi-process aware
So far the memory reporter aggregates reports from all processes, and runs the system reporter only in the main
process. Instead it is desirable to have per-process reports. We do so by:
- creating a ProcessReports struct that holds includes the pid in addition to the reports themselves.
- running the system memory reporter also in content processes.
- updating the about:memory page to create one report per process, and add useful information like the pid and
  the urls loaded in a given process.
