+++ components/constellation/Cargo.toml
@@ -39,6 +39,7 @@ media = { path = "../media" }
 net = { path = "../net" }
 net_traits = { workspace = true }
 parking_lot = { workspace = true }
+profile = { path = "../profile" }
 profile_traits = { workspace = true }
 script_layout_interface = { workspace = true }
 script_traits = { workspace = true }
