+++ components/script/script_thread.rs
@@ -71,7 +71,7 @@ use net_traits::{
     ResourceFetchTiming, ResourceThreads, ResourceTimingType,
 };
 use percent_encoding::percent_decode;
-use profile_traits::mem::ReportsChan;
+use profile_traits::mem::{ProcessReports, ReportsChan};
 use profile_traits::time::ProfilerCategory;
 use profile_traits::time_profile;
 use script_layout_interface::{
@@ -427,6 +427,7 @@ impl ScriptThreadFactory for ScriptThread {
                 memory_profiler_sender.run_with_memory_reporting(
                     || {
                         script_thread.start(CanGc::note());
+
                         let _ = script_thread
                             .senders
                             .content_process_shutdown_sender
@@ -2426,7 +2427,7 @@ impl ScriptThread {
             document.window().layout().collect_reports(&mut reports);
         }
 
-        reports_chan.send(reports);
+        reports_chan.send(ProcessReports::new(reports));
     }
 
     /// Updates iframe element after a change in visibility
